+++ apps/etl-func/src/domain/message.ts
@@ -117,6 +117,10 @@ export interface MessageRepository {
   ) => Promise<Message | undefined>;
 }
 
+export interface MessageEventProducer {
+  publishMessageEvent: (message: MessageEvent) => Promise<void>;
+}
+
 export class Message {
   content: MessageContent;
   id: string;
